+++ src/utils/misc.h
@@ -81,6 +81,7 @@ char dirSeparator();
 void waitMsecs(int msecs);
 QString portableDataPath();
 bool isInPortableMode();
+bool isAppImage();
 QString prependPortableDataPathIfNeeded(QString path,
                                         bool ifNotEmptyOnly = false);
 QString makePathRelativeToPortableDataPathIfNeeded(QString path);