+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   before showing the search result, so users have time to enter more search characters
   (for [#2335](https://github.com/pbek/QOwnNotes/pull/2335), thank you @fnkhz)
 - some database warnings were fixed when running the application for the first time
+- the data path detection was improved for AppImages in portable mode
+  (for [#2324](https://github.com/pbek/QOwnNotes/issues/2324#issuecomment-990642833))  
 
 ## 22.1.3
 - Joplin sub-folders can now also be imported if the parent folder is missing